Computer-Vision/Traffic-Sign-Classification/requirements.txt
Flask==2.2.2
numpy==1.23.4
tensorflow==2.10.0
Pillow==8.1.0
gunicorn